refactor(router): include currency conversion utility functions as Currency methods
