+++ crates/common_enums/Cargo.toml
@@ -12,6 +12,7 @@ dummy_connector = []
 
 [dependencies]
 diesel = { version = "2.1.0", features = ["postgres"] }
+error-stack = "0.3.1"
 serde  = { version = "1.0.160", features = [ "derive" ] }
 serde_json = "1.0.96"
 strum  = { version = "0.25", features = [ "derive" ] }
